# docs/openapi.yaml
# generated by tools/protocolgen — edit the declarations in sieve/protocol, not this file
openapi: 3.1.0
info:
  title: Sieve HTTP contract
  version: 1.0.0
  description: |-
    The typed JSON endpoints, generated by tools/protocolgen from sieve/protocol.
    Descriptions are the godoc comments and doc: struct tags on the declarations themselves.
    Routes serving hypermedia or bytes are deliberately absent: their contract is the
    template they render, and docs/API.md inventories them.
paths:
  /api/document/load:
    get:
      operationId: getApiDocumentLoad
      summary: DocumentLoadRequest reads a channel-less pseudo-document's content.
      description: |-
        DocumentLoadRequest reads a channel-less pseudo-document's content.

        It is the read half of the pair DocumentSaveRequest closes, and serves PROMPT
        PSEUDO-DOCUMENTS exclusively: a prompt never opens a document channel, so it
        has no wire to load along either. A note always has one and always loads
        through the wire's load frame — which is why this answers a prompt id only.
      parameters:
        - name: uuid
          in: query
          required: true
          description: the prompt pseudo-document id (prompt:{name}); a query parameter, not a body field
          schema:
            type: string
      responses:
        "200":
          description: a typed body, described by the entry's Response type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProtocolDocumentContent'
  /api/document/save:
    post:
      operationId: postApiDocumentSave
      summary: DocumentSaveRequest writes a document's buffer to disk.
      description: |-
        DocumentSaveRequest writes a document's buffer to disk.

        It is the write half of the channel-less pair, and serves PROMPT
        PSEUDO-DOCUMENTS exclusively: a prompt has no document channel, so it has no
        wire to flush along. A note always has one and always saves through it.
      parameters:
        - name: uuid
          in: query
          required: true
          description: the prompt pseudo-document id (prompt:{name}); a query parameter, not a body field
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                body:
                  type: string
                  description: the whole buffer
                mode:
                  type: string
                  description: records which mode saved it; empty leaves the recorded mode alone
              required:
                - body
      responses:
        "200":
          description: a typed body, described by the entry's Response type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProtocolDocumentSaveResponse'
  /api/folder:
    post:
      operationId: postApiFolder
      summary: FolderCreateRequest makes a new folder.
      description: FolderCreateRequest makes a new folder.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProtocolFolderCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProtocolFolderCreateRequest'
      responses:
        "200":
          description: HTML for HTMX to swap in; the template is its contract
          content:
            text/html:
              schema:
                type: string
  /api/folder/{id}:
    patch:
      operationId: patchApiFolderId
      summary: FolderPatchRequest changes a folder's properties.
      description: |-
        FolderPatchRequest changes a folder's properties. Both fields are POINTERS
        because this is a patch: absent must be distinguishable from empty-or-false, or
        renaming a folder would silently collapse it.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProtocolFolderPatchRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProtocolFolderPatchRequest'
      responses:
        "200":
          description: HTML for HTMX to swap in; the template is its contract
          content:
            text/html:
              schema:
                type: string
  /api/note/{id}:
    patch:
      operationId: patchApiNoteId
      summary: NotePatchRequest changes a note's properties.
      description: NotePatchRequest changes a note's properties.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProtocolNotePatchRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProtocolNotePatchRequest'
      responses:
        "200":
          description: HTML for HTMX to swap in; the template is its contract
          content:
            text/html:
              schema:
                type: string
  /api/tabs/close:
    post:
      operationId: postApiTabsClose
      summary: TabsCloseRequest closes every listed tab — the ONE close mechanism.
      description: |-
        TabsCloseRequest closes every listed tab — the ONE close mechanism. A single
        close sends one id, Close All sends every id, Close Others sends the complement
        of the kept tab: the client computes the set, the server closes it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProtocolTabsCloseRequest'
      responses:
        "200":
          description: HTML for HTMX to swap in; the template is its contract
          content:
            text/html:
              schema:
                type: string
components:
  schemas:
    BlockFrontendBlock:
      type: object
      description: |-
        FrontendBlock is the wire shape the WYSIWYG editor renders from (Stage D.2).
        It is a flattened, presentation-oriented projection of a SieveBlock with ONE
        uniform shape per kind: the payload always rides in Attrs (prose's body at
        Attrs["content"], structured props by key). The frontend renders prose to native
        nodes from Attrs["content"] and structured blocks through their renderer straight
        from Attrs — no markdown and no ProseMirror JSON cross the wire.
      properties:
        id:
          type: string
        kind:
          type: string
        attrs:
          type: object
        aliases:
          type: array
          items:
            type: string
      required:
        - id
        - kind
    ProtocolDocumentContent:
      type: object
      description: |-
        DocumentContent is a document as the editor mounts it: the buffer, the mode it
        was last in, and — in WYSIWYG — the block list it renders from.
      properties:
        body:
          type: string
          description: raw markdown, frontmatter already stripped
        mode:
          type: string
          description: wysiwyg | markdown — the mode this document was last in
        uuid:
          type: string
          description: empty when nothing was found, which the client treats as an empty document
        scroll:
          type: integer
          description: the tab's saved scroll offset; 0 parks at the top
        version:
          type: integer
          description: the version of the content served here; 0 when the container keeps no version history
        blocks:
          type: array
          description: 'WYSIWYG only: the block list, already through the shadow so ids are stable'
          items:
            $ref: '#/components/schemas/BlockFrontendBlock'
      required:
        - body
        - mode
        - uuid
        - scroll
        - version
    ProtocolDocumentSaveResponse:
      type: object
      description: |-
        DocumentSaveResponse reports the version the save produced. Its one caller
        saves a PROMPT, which the store writes as a plain file with no metadata, so
        the answer is always 0: this endpoint has no versioned container to describe.
        The saved-signal a client acts on is the container-saved fact, not this body.
      properties:
        version:
          type: integer
          description: the saved document's new version number; 0 for a container that keeps no version history
      required:
        - version
    ProtocolFolderCreateRequest:
      type: object
      description: FolderCreateRequest makes a new folder.
      properties:
        name:
          type: string
      required:
        - name
    ProtocolFolderPatchRequest:
      type: object
      description: |-
        FolderPatchRequest changes a folder's properties. Both fields are POINTERS
        because this is a patch: absent must be distinguishable from empty-or-false, or
        renaming a folder would silently collapse it.
      properties:
        name:
          type: string
          description: the new display name; absent leaves it alone
        open:
          type: boolean
          description: whether the folder is expanded in the tree; absent leaves it alone
    ProtocolNotePatchRequest:
      type: object
      description: NotePatchRequest changes a note's properties.
      properties:
        name:
          type: string
          description: the new display name
      required:
        - name
    ProtocolTabsCloseRequest:
      type: object
      description: |-
        TabsCloseRequest closes every listed tab — the ONE close mechanism. A single
        close sends one id, Close All sends every id, Close Others sends the complement
        of the kept tab: the client computes the set, the server closes it.
      properties:
        ids:
          type: array
          description: the tabs to close
          items:
            type: string
      required:
        - ids